change for short compact

+++ src/helpers/format-big-number.ts
@@ -1,3 +1,3 @@
 export function formatBigNumber(number: number): string {
-  return new Intl.NumberFormat('fr-FR', { notation: 'compact', compactDisplay: 'long' }).format(number);
+  return new Intl.NumberFormat('fr-FR', { notation: 'compact', compactDisplay: 'short' }).format(number);
 }
